Remove row id from QueuedItem class.
Row id in QueuedItem class was only used to sort dirty items by
their row ids with sqlite as backend.

Reviewed-on: http://review.couchbase.org/20238
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>

+++ src/checkpoint.cc
@@ -1117,7 +1117,7 @@ queued_item CheckpointManager::createCheckpointItem(uint64_t id,
     } else {
         key << "checkpoint_end";
     }
-    queued_item qi(new QueuedItem(key.str(), vbid, checkpoint_op, (int64_t) id));
+    queued_item qi(new QueuedItem(key.str(), vbid, checkpoint_op, id));
     return qi;
 }
